% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mici.impute.R
\name{mici.impute}
\alias{mici.impute}
\title{Function to generate multiple imputations from the mici package by Elizabeth Chase}
\usage{
mici.impute(ftime = NULL, ftype = NULL, data = NULL, M = 200, bootstrap)
}
\arguments{
\item{ftime}{Either the name of the variable in the data that contains the event
times, or a length n vector of event times.}

\item{ftype}{Either the name of the variable in the data that contains the event
types, or a length n vector of event types. Note that event = 0 corresponds to
censoring, event = 1 corresponds to the event of interest, and any event indicator
greater than 1 corresponds to competing event(s).}

\item{data}{Optional: if the ftime and ftype inputs are variable names (rather than
numeric vectors), this is the dataframe in which those variables are stored.}

\item{M}{The number of imputations desired; the default is 200.}
}
\value{
The function returns a length M list, where each object of the list
is a dataframe with columns ftime, ftype containing the imputed event times and
event types. If data was inputted to the original function, all original columns
of the dataset are kept in each of the M dataframes of the list. If only vectors
were inputted to the original function, the columns time and type in each of the
M dataframes store the original inputs (without imputation).
}
\description{
#' The source code for mici.impute is included in the msmi package because
mici is not currently available on CRAN.
From the authors of the mici package,
"This function generates multiple imputations according to the nonparametric
multiple imputation scheme described in Chase et al. (2025). In particular,
for censored individuals it generates an imputed event time and type by sampling
from the Kaplan-Meier distribution of overall event times and then randomly
sampling an event type among those observed at the imputed event time. The
function outputs a length M list of imputed datasets free of censoring."
}
