% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmi.plot.R
\name{msmi.plot}
\alias{msmi.plot}
\title{Plot Confidence Regions State Occupation Probabilities}
\usage{
msmi.plot(tprob_obj, time)
}
\arguments{
\item{tprob_obj}{The list returned by msmi.tprobs()}

\item{time}{The time point of interest at which to generate the plot}
}
\value{
A plot showing the confidence region and point estimate in the probability simplex.
}
\description{
Plot Confidence Regions State Occupation Probabilities
}
\examples{
imps <- msmi.impute(dat = sim.data, M = 5,
                    prefix.states = c("event", "t"), method = "marginal")
tprobs <- msmi.tprobs(imp_obj = imps, times = c(1,3))
#Plot the confidence region for state occupation probabilities at time = 1
msmi.plot(tprob_obj = tprobs, time = 1)
}
