% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmi.tprobs.R
\name{msmi.tprobs}
\alias{msmi.tprobs}
\title{Calculate Empirical Transition Matrix from Multiple Imputed Datasets for an Illness-Death Model}
\usage{
msmi.tprobs(
  imp_obj = NULL,
  times = NULL,
  method = "trWald",
  alpha = 0.05,
  vcov = FALSE
)
}
\arguments{
\item{imp_obj}{Output from msmi.impute()}

\item{times}{Vector of times at which to calculate state occupation probabilities}

\item{method}{The interval estimation method to use: "trWald" or "goodman" (default is "trWald")}

\item{alpha}{Significance level for the confidence region (default is 0.05)}

\item{vcov}{Logical indicating whether to return the variance-covariance matrices for the state occupation probabilities at each time (default is FALSE)}
}
\value{
A list with the following components:
\item{mi_estimate}{A data frame with the multiple imputation point estimates
of state occupation probabilities at each time}
\item{method}{The method used for confidence region construction}
\item{alpha}{The significance level for the confidence region}
\item{vcov}{(if vcov = TRUE and method = "trWald") A list of variance-covariance matrices for the state occupation probabilities at each time in the unconstrained space}
\item{cr_list}{A list of confidence regions (defined via convex hulls) for the state occupation probabilities at each time
in the probability space.}
}
\description{
Calculate Empirical Transition Matrix from Multiple Imputed Datasets for an Illness-Death Model
}
\examples{
imps <- msmi.impute(dat = sim.data, M = 5,
            prefix.states = c("event", "t"), method = "marginal")
msmi.tprobs(imp_obj = imps, times = seq(1, 5, 1))
}
