% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmi.tprobs.v2.R
\name{msmi.tprobs.v2}
\alias{msmi.tprobs.v2}
\title{Calculate Empirical Transition Matrix from Multiple Imputed Datasets for an Illness-Death Model}
\usage{
msmi.tprobs.v2(imp_obj = NULL, times = NULL, method = "trWald", alpha = 0.05)
}
\arguments{
\item{imp_obj}{Output from msmi.impute()}

\item{times}{Vector of times at which to calculate state occupation probabilities}

\item{method}{The interval estimation method to use: "trWald" or "goodman" (default is "trWald")}

\item{alpha}{Significance level for the confidence region (default is 0.05)}
}
\value{
A list with the following components:
\item{mi_estimate}{A data frame with the multiple imputation point estimates
of state occupation probabilities at each time}
\item{method}{The method used for confidence region construction}
\item{alpha}{The significance level for the confidence region}
\item{cr_list}{A list of confidence regions (defined via convex hulls) for the state occupation probabilities at each time
in the probability space.}
}
\description{
Calculate Empirical Transition Matrix from Multiple Imputed Datasets for an Illness-Death Model
}
\examples{
imps <- msmi.impute(dat = sim.data, M = 5,
            prefix.states = c("event", "t"), method = "marginal")
msmi.tprobs.v2(imp_obj = imps, times = seq(1, 5, 1))
}
