% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmi.nested.impute.R
\name{nested_marginal_mi}
\alias{nested_marginal_mi}
\title{Layer 2 Imputation: Marginal Approach}
\usage{
nested_marginal_mi(d, m, r)
}
\arguments{
\item{d}{A data frame with one row per subject and the columns event1, t1, event2, t2 on which mici::mici.impute was previously ran}

\item{m}{An integer, the nest id for the nested imputation structure}

\item{r}{An integer, the impute id for the nested imputation structure}
}
\value{
A data frame with imputed times for event2 where event2 was censored
}
\description{
Layer 2 Imputation: Marginal Approach
}
