% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.region.R
\name{plot_region}
\alias{plot_region}
\title{Plot Confidence Regions State Occupation Probabilities}
\usage{
plot_region(tprob_obj, time, unconstrained = FALSE)
}
\arguments{
\item{tprob_obj}{The list returned by msmi::tprobs()}

\item{time}{The time point of interest at which to generate the plot}

\item{unconstrained}{Logical indicating whether to provide an additional plot in the unconstrained 2D space. Default is FALSE.}
}
\value{
If unconstrained == FALSE, A plot showing the confidence region and point estimate in the probability simplex.
If unconstrained == TRUE, A list of plots showing the confidence region and point estimate in the (1) unconstrained 2D space and the (2) probability simplex.
}
\description{
Plot Confidence Regions State Occupation Probabilities
}
\examples{
imps <- msmi.impute(dat = sim.data, M = 5, n.states = 3,
                    prefix.states = c("event", "t"), method = "marginal")
tprobs <- msmi.tprobs(imp_obj = imps, times = c(1,3))
#Plot the confidence region for state occupation probabilities at time = 1
plot_region(tprob_obj = tprobs, time = 1)
}
