% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmi.tprobs.R, R/msmi.tprobs.v2.R
\name{get_empirical_probs}
\alias{get_empirical_probs}
\title{Calculate empirical state occupation probabilities for a single imputed dataset}
\usage{
get_empirical_probs(df, times)

get_empirical_probs(df, times)
}
\arguments{
\item{df}{A data frame with imputed event times and event indicators}

\item{times}{Vector of times at which to calculate state occupation probabilities}
}
\value{
A data frame with empirical state occupation probabilities at specified times

A data frame with empirical state occupation probabilities at specified times
}
\description{
Calculate empirical state occupation probabilities for a single imputed dataset

Calculate empirical state occupation probabilities for a single imputed dataset
}
