% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmi.impute.R
\name{marginal_mi}
\alias{marginal_mi}
\title{Layer 2 Imputation: Marginal Approach}
\usage{
marginal_mi(d)
}
\arguments{
\item{d}{A data frame with one row per subject and the columns event1, t1, event2, t2 on which mici::mici.impute was previously ran}
}
\value{
A data frame with imputed times for event2 where event2 was censored
}
\description{
Layer 2 Imputation: Marginal Approach
}
