% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{msmi.impute}
\alias{msmi.impute}
\title{Create multiple imputed datasets for data arising from multi-state models subject to censoring}
\usage{
msmi.impute(
  dat,
  M,
  n.states = 3,
  prefix.states = c("event", "t"),
  method = "marginal",
  seed = sample(1:.Machine$integer.max, size = 1)
)
}
\arguments{
\item{dat}{a dataframe with one row per subject and columns corresponding to the time and event indicators for each state transition.
The column names for the time and event indicators should be in the format specified by prefix.states, specifically
event\if{html}{\out{<i>}} and t\if{html}{\out{<i>}} for i = 1, ..., n.states-1}

\item{M}{an integer, the number of imputations}

\item{n.states}{an integer, the number of states in the multistate model}

\item{prefix.states}{a character vector of length 2, specify the prefix for the event and time columns in the d in that order}

\item{method}{a character string, either "marginal" or "cox", indicating which method to use for the second layer of imputation}

\item{seed}{an integer, the seed for random number generation}
}
\value{
A list of length M, where each element is a data frame with imputed times for censored events
}
\description{
Create multiple imputed datasets for data arising from multi-state models subject to censoring
}
\examples{
msmi.impute(sim.data, M = 5, n.states = 3, prefix.states = c("event", "t"), method = "marginal")
}
