% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmi.tprobs.R
\name{msmi.tprobs}
\alias{msmi.tprobs}
\title{Calculate Empirical Transition Matrix from Multiple Imputed Datasets for an Illness-Death Model}
\usage{
msmi.tprobs(
  imp_obj = NULL,
  times = NULL,
  int.type = "multinomial_logistic",
  alpha = 0.95,
  vcov = FALSE
)
}
\arguments{
\item{imp_obj}{Output from msmi.impute()}

\item{times}{Vector of times at which to calculate state occupation probabilities}

\item{int.type}{Type of confidence region to compute: "multinomial_logistic" or "dirichlet" (default is "multinomial_logistic")}

\item{alpha}{Significance level for the confidence region (default is 0.95)}

\item{vcov}{Logical indicating whether to return the variance-covariance matrices for the state occupation probabilities at each time (default is FALSE)}
}
\value{
A list with the following components:
\item{mi_estimate}{A data frame with the multiple imputation point estimates
of state occupation probabilities at each time}
\item{int.type}{The type of confidence interval used}
\item{alpha}{The significance level for the confidence region}
\item{vcov}{if vcov = TRUE, A list of variance-covariance matrices for the state occupation probabilities at each time in the unconstrained space}
\item{cr_list}{A list of confidence regions (defined via point clouds) for the state occupation probabilities at each time
in the (1) unconstrained 2D space and (2) the probability space.}
}
\description{
Calculate Empirical Transition Matrix from Multiple Imputed Datasets for an Illness-Death Model
}
\examples{
imps <- msmi.impute(dat = sim.data, M = 5, n.states = 3,
            prefix.states = c("event", "t"), method = "marginal")
msmi.tprobs(imp_obj = imps, times = seq(1, 5, 1))
}
